BL-12495 code review move invisibles replacement to own file
